+++ package.json
@@ -5,17 +5,17 @@
     "main": "index.js",
     "type": "module",
     "scripts": {
-        "start": "node ./index.js",
-        "start:dev": "nodemon ./index.js",
-        "test": "echo \"Error: no test specified\" && exit 1"
+      "start": "tsx ./index.ts",
+      "start:dev": "tsx watch ./index.ts",
+      "test": "echo \"Error: no test specified\" && exit 1"
     },
     "author": "Andrei Auchynnikau",
     "keywords": [],
     "license": "ISC",
-    "devDependencies": {
-        "nodemon": "^2.0.16"
-    },
     "dependencies": {
         "ws": "^8.8.0"
+    },
+    "devDependencies": {
+        "tsx": "^4.7.1"
     }
 }
